Backport some minor changes from int-new
